requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
boto3==1.28.85
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
cryptography==41.0.7
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0